backend/requirements.txt
asgiref==3.8.1
certifi==2023.7.22
charset-normalizer==3.3.2
distlib==0.3.7
Django==4.2.11
djangorestframework==3.15.1
feedparser==6.0.11
filelock==3.12.4
idna==3.6
importlib-metadata==7.1.0
joblib==1.3.2
lightgbm==4.1.0
Markdown==3.6
numpy==1.26.1
pipenv==2023.10.3
platformdirs==3.11.0
pyshorteners==1.0.1
python-dotenv==1.0.0
requests==2.31.0
scikit-learn==1.3.2
scipy==1.11.3
sgmllib3k==1.0.0
sqlparse==0.4.4
threadpoolctl==3.2.0
typing-extensions==4.10.0
urllib3==2.2.1
virtualenv==20.24.5
zipp==3.18.1
